+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- `@LocalNode` annotation to receive the local Node
+
 ## 0.1.0 - 2015-03-05
 ### Added
 - Initial release