+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.2.0 - 2015-03-05
 ### Added
 - `@LocalNode` annotation to receive the local `Node`
 - `@RemoteNodes` annotation to receive a list of connected `Node`s